requirements.txt
streamlit
transformers
torch
bitsandbytes
accelerate
Pillow
huggingface-hub
uuid
numpy
sentence-transformers
faiss-gpu